fix: address CodeRabbit review on Reloop .NET v2
Scope auth headers per request, preserve null clears in update bodies,
encode domain path IDs, tighten webhook/inbox validation, and align
property-key patterns.

+++ Models/WebhookModels.cs
@@ -124,10 +124,10 @@ public static class WebhookModels
         public bool? RateLimitEnabled { get; set; }
 
         [JsonPropertyName("maxRequestsPerMinute")]
-        public double? MaxRequestsPerMinute { get; set; }
+        public int? MaxRequestsPerMinute { get; set; }
 
         [JsonPropertyName("maxRetries")]
-        public double? MaxRetries { get; set; }
+        public int? MaxRetries { get; set; }
 
         [JsonPropertyName("retryBackoffMultiplier")]
         public double? RetryBackoffMultiplier { get; set; }
